LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/fill-and-stroke-styles/2d.gradient.conic.negative.rotation-expected.txt

2d.gradient.conic.negative.rotation
Conic gradient with negative rotation
Actual output:
Expected output:


FAIL Conic gradient with negative rotation assert_equals: Red channel of the pixel at (25, 15) expected 255 but got 195
